renamed happiness to HappinessManager and added distance to light hapinesss measure
